ndctl/namespace.c: fix a potential integer overflow
Static analysis warns that we may cause an integer overflow when
left-shifting bb->len in nstype_clear_badblocks(). While it is unlikely
to trigger this in real-world usage, since bb->len is limited by the
kernel, the fix is is a trivial cast, and it quells the warning, and
increases our robustness.

+++ ndctl/namespace.c
@@ -1146,7 +1146,7 @@ static int nstype_clear_badblocks(struct ndctl_namespace *ndns,
 		unsigned long long bb_begin, bb_end, bb_len;
 
 		bb_begin = region_begin + (bb->offset << 9);
-		bb_len = bb->len << 9;
+		bb_len = (unsigned long long)bb->len << 9;
 		bb_end = bb_begin + bb_len - 1;
 
 		/* bb is not fully contained in the usable area */
